Fix generated text-files
Text files generated (from redirect templates and versions.json) are
missing trailing newlines ("No newline at end of file").

Missing in the sense that `git diff` and similar operations are emitting
such diagnostic messages.

The issue is with these kinds of files:

- Redirect html files generated from the redirect template
- versions.json

Fix is to add a newline (U000A LINE FEED (LF) \n) at the end of file. It is
verbatim as git stores text-files with the \n line separator internally
and these buffers are used as file-data in `git-fast-import`.